roles: enforce "at least one coinbase output is present" at config parsing time
Saves us a bunch of run-time asserts/unwraps/expects.

The existing example configuration files say "Right now only one output
is supported, so comment all the ones you don't need !". But this
condition is not actually enforced anywhere. It's just that when push
comes to shove, we only actually use the first entry of the vector.

So maybe this commit should enforce that exactly one entry is present?
Right now it only enforces that *at least* one entry is present, which
matches the existing behavior.

The configuration structs all have `new` constructors which I believe
are only public so that they can be used in the integration tests.
Therefore I simply made them panic if you try to use them to construct a
config object with a bad number of coinbase outputs.